Refactor : sellerDto 중복 제거
