+++ src/main/java/backend/mossy/shared/member/event/SellerJoinedEvent.java
@@ -1,6 +1,6 @@
 package backend.mossy.shared.member.event;
 
-import backend.mossy.shared.payout.dto.event.SellerDto;
+import backend.mossy.shared.member.dto.event.SellerDto;
 
 public record SellerJoinedEvent(
         SellerDto seller
