+++ src/main/java/backend/mossy/shared/payout/event/PayoutSellerCreatedEvent.java
@@ -1,6 +1,6 @@
 package backend.mossy.shared.payout.event;
 
-import backend.mossy.shared.payout.dto.event.SellerDto;
+import backend.mossy.shared.member.dto.event.SellerDto;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
